+++ superset/assets/src/visualizations/TreeMap/index.js
@@ -1,0 +1,6 @@
+import adaptor from './adaptor';
+import treemap from './treemap';
+
+export { treemap };
+export default adaptor;
+
